Remove Repository.parse_diff, remove repo check from Diff.merge
